REF: refactor cumulative op tests from test_analytics
